24_03_18
Item Manager 스크립트 50% 완성
Item Inventory 테스트 해봐야함
Ore Inventory 체력 부분 누락 사항 추가, 능력치 추가 부분 변경

+++ ProjectUDF/Assets/01. Scripts/gusdnr/Inventory/Ore/MonoOre.cs
@@ -3,7 +3,7 @@ using UnityEngine;
 public class MonoOre : MonoBehaviour
 {
 	public Stats IncreaseStat;
-	public float IncreaseValue = 3;
+	public int IncreaseValue = 3;
 
 	public void GetOre() => OreInventory.Instance.AddOre(IncreaseStat, IncreaseValue);
 }
